requirements.txt
fastapi==0.117.1
passlib==1.7.4
pydantic==2.11.9
pydantic_settings==2.11.0
PyJWT==2.10.1
PyJWT==2.10.1
python_jose==3.3.0
Requests==2.32.5
alembic
uvicorn
psycopg[binary]>=3.1,<4.0 # Библиотека для работы с PostgreSQL
jinja2
python-multipart
httpx
redis
